My hours home - today

+++ plugins/fmcWorkingHourPlugin/modules/workingHourUser/templates/_dayitems_list.php
@@ -25,17 +25,18 @@
             <?php echo $item->getComment(); ?>
         </td>
         
-        <td>
-            <?php #$editurl = url_for('workingHourUser_edititem', array('date' => $date,'item_id' => $item->getId())); ?>
-            <a href="<?php echo $editurl; ?>">
-                <i class="icon-pencil"></i>
-            </a> 
-            
-            <?php #$deleteurl = url_for('workingHourUser_deleteitem', array('item_id'=>$item->getId())); ?>
-            <a href="<?php echo $deleteurl; ?>">
-                <i class="icon-trash"></i>
-            </a>
-        </td>
+        <?php if ($edit): ?>        
+            <td>
+                <a href="<?php echo $editurl; ?>">
+                    <i class="icon-pencil"></i>
+                </a> 
+                
+                <a href="<?php echo $deleteurl; ?>">
+                    <i class="icon-trash"></i>
+                </a>
+            </td>
+        <?php endif; ?>
+        
     </tr>
     
 <?php endforeach; ?>
